backend/app/requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
pandas==2.3.1
numpy==1.26.4
torch==2.7.0
transformers==4.36.0
scikit-learn==1.6.1
pymongo==4.11.2
python-dotenv
langchain==0.3.27
langchain_groq==0.3.7